# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sistema de Gestão de Projetos API",
    "description": "API para gerenciar o ciclo de vida de projetos.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/projects/{project_id}/tasks/": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "summary": "Lista Tarefas do Projeto",
        "description": "Retorna todas as tarefas associadas a um projeto específico. Requer autenticação de qualquer usuário válido.",
        "operationId": "get_tasks_for_project_api_projects__project_id__tasks__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskRead"
                  },
                  "title": "Response Get Tasks For Project Api Projects  Project Id  Tasks  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tasks"
        ],
        "summary": "Cria Nova Tarefa",
        "description": "Cria uma nova tarefa dentro de um projeto específico. Envia notificação automática se um usuário for atribuído à tarefa. Requer autenticação de qualquer usuário válido.",
        "operationId": "create_task_api_projects__project_id__tasks__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/tasks/{task_id}": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "summary": "Busca Tarefa por ID",
        "description": "Retorna os detalhes de uma tarefa específica dentro de um projeto. Valida se a tarefa pertence ao projeto informado. Requer autenticação de qualquer usuário válido.",
        "operationId": "get_task_api_projects__project_id__tasks__task_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Tasks"
        ],
        "summary": "Atualiza Tarefa Existente",
        "description": "Atualiza os dados de uma tarefa existente dentro de um projeto. Envia notificação automática se houver mudança na atribuição de usuário. Requer autenticação de qualquer usuário válido.",
        "operationId": "update_task_api_projects__project_id__tasks__task_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Tasks"
        ],
        "summary": "Exclui uma Tarefa",
        "description": "Remove permanentemente uma tarefa de um projeto. Valida se a tarefa pertence ao projeto informado. Requer permissão de MANAGER ou ADMIN.",
        "operationId": "delete_task_api_projects__project_id__tasks__task_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TaskCreate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Título descritivo da tarefa",
            "example": "Configurar Active Directory"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Descrição detalhada da tarefa",
            "example": "Configurar estrutura organizacional do AD com OUs e grupos de segurança"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus",
            "description": "Status atual da tarefa",
            "default": "todo",
            "example": "TODO"
          },
          "priority": {
            "$ref": "#/components/schemas/TaskPriority",
            "description": "Prioridade da tarefa",
            "default": "medium",
            "example": "HIGH"
          },
          "dueDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duedate",
            "description": "Data limite para conclusão da tarefa",
            "example": "2025-02-15"
          },
          "assigned_to_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To Id",
            "description": "ID do usuário responsável pela tarefa",
            "example": "550e8400-e29b-41d4-a716-446655440007"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "TaskCreate"
      },
      "TaskPriority": {
        "type": "string",
        "enum": [
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "TaskPriority"
      },
      "TaskRead": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Título descritivo da tarefa",
            "example": "Configurar Active Directory"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Descrição detalhada da tarefa",
            "example": "Configurar estrutura organizacional do AD com OUs e grupos de segurança"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus",
            "description": "Status atual da tarefa",
            "default": "todo",
            "example": "TODO"
          },
          "priority": {
            "$ref": "#/components/schemas/TaskPriority",
            "description": "Prioridade da tarefa",
            "default": "medium",
            "example": "HIGH"
          },
          "dueDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duedate",
            "description": "Data limite para conclusão da tarefa",
            "example": "2025-02-15"
          },
          "assigned_to_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To Id",
            "description": "ID do usuário responsável pela tarefa",
            "example": "550e8400-e29b-41d4-a716-446655440007"
          },
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Identificador único da tarefa",
            "example": "550e8400-e29b-41d4-a716-446655440008"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id",
            "description": "ID do projeto ao qual a tarefa pertence",
            "example": "550e8400-e29b-41d4-a716-446655440003"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat",
            "description": "Data e hora de criação da tarefa",
            "example": "2025-01-15T10:30:00Z"
          },
          "assigned_to": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserInTask"
              },
              {
                "type": "null"
              }
            ],
            "description": "Dados do usuário responsável pela tarefa"
          }
        },
        "type": "object",
        "required": [
          "title",
          "id",
          "project_id",
          "createdAt"
        ],
        "title": "TaskRead"
      },
      "TaskStatus": {
        "type": "string",
        "enum": [
          "todo",
          "in-progress",
          "done",
          "hold"
        ],
        "title": "TaskStatus"
      },
      "TaskUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "description": "Novo título da tarefa",
            "example": "Configurar Active Directory - Atualizado"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Nova descrição da tarefa",
            "example": "Configurar AD com integração ao Azure AD"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskStatus"
              },
              {
                "type": "null"
              }
            ],
            "description": "Novo status da tarefa",
            "example": "IN_PROGRESS"
          },
          "priority": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskPriority"
              },
              {
                "type": "null"
              }
            ],
            "description": "Nova prioridade da tarefa",
            "example": "URGENT"
          },
          "dueDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duedate",
            "description": "Nova data limite",
            "example": "2025-02-20"
          },
          "assigned_to_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To Id",
            "description": "Novo responsável pela tarefa",
            "example": "550e8400-e29b-41d4-a716-446655440009"
          }
        },
        "type": "object",
        "title": "TaskUpdate"
      },
...
      "UserInTask": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Identificador único do usuário",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "email": {
            "type": "string",
            "title": "Email",
            "description": "Endereço de email do usuário",
            "example": "maria.santos@empresa.com"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email"
        ],
        "title": "UserInTask"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/token"
          }
        }
      }
    }
  }
}